+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Update copyright notice.
 * Update dependencies.
 
+### Fixed
+
+* Escape regex patterns used for variable renaming.
+
 ## [0.6.6] - 2025-12-06
 
 ### Changed